refactor(frontend): lift App.tsx state into per-feature controller hooks
Point 3.5 of #103: complete the App.tsx split by moving state, refs,
effects, and handlers into per-feature controller hooks. App.tsx
2686 → 271 lines (90% reduction); the shell now only owns viewMode +
auth mode, composes 6 controllers, wires cross-feature side effects,
and renders the right feature view.

New controllers under frontend/src/features/<name>/:
- auth/useAuthController.ts (149 lines) — current-user query + login/
  register/logout mutations + form validation. Returns formProps + a
  logout fn. Caller passes an `onLoginSuccess` callback so the shell
  can clear the gallery cache without the hook touching unrelated
  state.
- folders/useFoldersController.ts (453) — folder list + uploads +
  scan polling (5s setInterval kept; refetchInterval refactor is
  follow-up). Exposes `refreshFolders`, `panelProps`,
  `folderDetailsById`, `orderedFolders`. Caller passes
  `onUploadComplete` + `onScanFinished` callbacks so the gallery
  can reload after either.
- duplicates/useDuplicatesController.ts (386) — scan settings, scan
  status, groups, action state. Polling uses an imperative loop
  (1:1 with App.tsx) — refetchInterval refactor is follow-up.
- favorites-sauce/useSauceFavoritesController.ts (475) — sauces,
  favorites sync (refetchInterval based on `status === 'running'`,
  replacing the manual `setInterval` + `favoritesPollRef`), legacy
  credentials, booru dev-options toggle. Exposes `settingsProps`,
  `sauceSettings`, `credentialsReady`, `favoritesSyncStatus`.
- library/useGalleryController.ts (702) — gallery state + the entire
  hand-rolled fetch pipeline (cache, AbortController, offset
  accumulator, manual order DnD). Preserved verbatim; useInfiniteQuery
  refactor is follow-up. Exposes `viewProps`, `goRelative`, `openFile`,
  `resetGallery`, `reloadGallery`, plus cross-feature helpers for
  file-detail (`selectedFileIndex`, `updateFavoriteFlag`,
  `removeFileFromGallery`, `savedGalleryScrollRef`).
- file-detail/useFileDetailController.ts (1137) — selectedFile,
  providers, tags, swipe carousel, fullscreen, all per-file
  mutations. Accepts `{ gallery, sauceSettings }` as input so it
  doesn't reach into other features' state directly.

App.tsx (now 271 lines):
- Shell-level state: `viewMode`, plus the small refs needed to
  break circular controller dependencies (`selectedFileRef`,
  `openFileRef`, `galleryCtlRef`).
- Cross-feature wiring: `goRelativeWrapper` (file-detail asks gallery
  to advance + load-more), `onFileOpen` override on gallery viewProps
  (also calls file-detail openFile), auth-required event handler
  (clears query cache across all domains via `queryClient.removeQueries`).
- Render: AuthForm gate → file-detail full-screen → settings/duplicates
  /gallery branch by viewMode.

Behavior preserved end-to-end: vitest 13/13, Playwright smoke 4/4,
vite build green. Bundle: CSS 79KB / JS 348KB (gz 13/99), essentially
unchanged from point 3.

Refs #103
